Characters automatically reload gasmasks when low
I recently visited a collapsed tower, and as I fought my way through the
dungeon, I faced a consistent issue: my gas mask would run out, and I
would not notice, becoming poisoned.

This was, needless to say, quite annoying. So, let's have the character
manage the changing of filters, instead of the player.

The solution is simple. Add a function to Character that will will take
an item and look at it's remaining ammo. If the item is low, but not
quite out, the function will warn every 8 turns about the item being
low. When the item runs out, it will scan the player inventory and
reload with appropriate items.

Then, this function simply needs to be called wherever the item consumes
charges, and the player will automatically handle reloading.

A variety of behavior tweaks are available, but by default (and always,
as there is no way to change them), a popup shown when the last suitable
item has been reloaded (next reload will fail) and when a reload is
attempted but fails.

Several tests have been added to ensure this behavior works for both
players and npcs.