Support specifying item vitamins in weight
Often, nutrition info will list vitamins by weight. Additionally, RDA,
being defined by humans, is subject to change. For these reasons, it
should be possible to specify vitamins by weight in items definitions.

To do this, vitamins need to know how weight corresponds to units, and
items need to load in weight as well as units, and convert to units as
necessary. Unfortunately, as items may be loaded before vitamins, both
the loaded weight values and units values must be stored until all data
is loaded and they can be converted to units. To facilitate this, turn
the vitamins data stored in nutrients into a map from vitamin id to a
union of int and units::mass. This can then be finalized after loading
to convert all units::mass values to units.

Several functions were added to ease the difficulty in navigating this
new structure, and safety checks were added to prevent bugs.

LIMITATIONS:
There's a lot of rounding that the whole vitamins system is subject to.
This adds some more rounding: Any fractional vitamins due to weight are
discarded. However, this is the same as the previous approach, but
happens at load time instead of when the content is being added.
Additionally, vitamins (like iron) that have an RDA of less than 1mg
cannot really benefit from this system.