Demonstrate translation fallback
This shows all the places that need to be touched to implement
translation fallback. Actual implementation requires implementing the
caching, as well as handling plurals and context.

To test, spawn in the untranslated moisturizer variant, as well as
another moisturizer variant. In English, you have untranslated
moisturizer and the other variant. When you switch to Spanish, you have
the translated name of the variant and generic moisturizer.